add more features

+++ WonderfulJourney/routes/web.php
@@ -15,34 +15,34 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
-Route::view('/', 'HomePage')->name('HomePage');
+Route::view('/', 'HomePage')->name('home');
+Route::get('/category/{id}/articles', function ($id) {
+    $category = Category::find($id);
+    return view('viewArticlesByCategory',compact('category'));
+})->name('ShowArticleByCategory');
 
 Auth::routes();
 
-Route::middleware(['auth', 'admin'])->group(function () {
-    Route::get('/admins', 'AdminController@index')->name('ViewAdminUsers');
-    Route::get('/members', 'UserController@index')->name('ViewMemberUsers');
+Route::middleware(['auth','admin'])->group(function () {
+    Route::get('/admins', 'UserController@showAdminUsers')->name('ViewAdminUsers');
+    Route::get('/members', 'UserController@showAdminMembers')->name('ViewMemberUsers');
 });
 
-Route::group(['prefix' => 'user'], function () {
-    Route::get('/', 'UserController@edit')->name('ShowUpdateUserForm');
-    Route::post('/{user}', 'UserController@update')->name('SubmitUpdateUser');
-    Route::delete('/{user}', 'UserController@destroy')->name('DeleteUser');
-    Route::get('/{user}/articles', 'UserController@getArticles')->name('ShowMemberArticles');
+Route::group(['prefix' => 'user','middleware' => ['auth','member']], function () {
+    Route::get('/edit/{user}', 'UserController@edit')->name('ShowUpdateUserForm');
+    Route::put('/edit/submit/{user}', 'UserController@update')->name('SubmitUpdateUser');
+    Route::delete('/{user}', 'UserController@destroy')->name('DeleteUser')->withoutMiddleware('member')->middleware('admin');
+    Route::get('/articles/{user}', 'UserController@getArticles')->name('ShowMemberArticles');
 
 });
 
-Route::group(['prefix' => 'article'], function () {
-    Route::delete('{article}', 'ArticleController@index')->name('DeleteArticle');
-    Route::post('/{article}', 'ArticleController@store')->name('SubmitPostArticle');
+Route::group(['prefix' => 'article','middleware' => ['auth','member']], function () {
+    Route::delete('{article}', 'ArticleController@destroy')->name('DeleteArticle')->withoutMiddleware('member')->middleware('admin');
+    Route::post('/u/submit', 'ArticleController@store')->name('SubmitPostArticle');
     Route::get('/article/form', 'ArticleController@create')->name('ShowPostArticleForm');
-    Route::get('/{id}/detail', 'ArticleController@show')->name('ShowArticleDetail');
+    Route::get('/{article}/detail', 'ArticleController@show')->name('ShowArticleDetail')->withoutMiddleware(['auth','member']);
 });
 
-Route::get('/category/{id}/articles', function ($id) {
-    $category = Category::find($id)->get();
-    return view('viewArticlesByCategory',compact('category'));
-})->name('ShowArticleByCategory');
